FIX: Reorganize files and resolve build issues
This commit reorganize source files so it can be build with Docker.

First issue that appeared was with Python, where we didn't have a
requirements.txt file that guides what dependencies are needed in
order to build the source files. I have created one based on try and
error.

Second issue with the Python source it was throwing errors during the
startup. After digging a little bit, it seems that the python-protobuf
library that I installed was newer than the code. Additionally, the code
is auto-generated by the protobuf binary. Then I just rebuild it on the
fly using a build container on Python's Dockerfile. And for Python, I
also created a requirements.txt.

Finally, the Golang code had external dependencies that are in a
private Github repository. Fortunately, the code was a fork from a
public Github repository. Instead of using /sybogames/ as source, I have
changed it to Google's gRPC examples.