NEW: Create diagram with the source and imgs
This commit adds the diagrams that shows how the app behaves in terms
of network traffic on AWS and within the K8s cluster.

Along with it, I am adding the source code of the diagram and the
requirements.txt that is needed to run this source code to either
modify the images or generate them again if needed.