FIX: Make some information follows the values.yaml
This commit is a minor fix into the NOTES.txt, where some values where
not exactly as shown when running some set of commands.

For instance, the hostname in the ingress resource is not pointing
towards another host address and the NOTES.txt wasn't following this new
name. With this fix it will.